Fixed Docker Copy Error
The Docker builds out a python:3.11 as builder environment and was copying that over the uemarketplace directly removing all the directories copied files.  Additionally there were path issues and problems with CRLF unix returns on the end of the line.
This has all been corrected and tested locally and is now working.